Format NHGIS places data by joining additional attributes and adding a label column

fmt_nhgis_places() joins Census division, region, and USPS abbreviation columns based on the state value and creates a new label column following the pattern of ", ".

Usage

fmt_nhgis_places(
  data,
  label_col = "label",
  name_col = "NAME",
  remove_pattern = " city$"
)

Arguments

label_col

Label column name.

name_col

Name column to use as the basis for the label.

remove_pattern

Passed to pattern for stringr::str_remove(). Defaults to " city$" to trim repetitive text from the end of the place names.