Format a percent value column with rounding and handling for 0 and NA values

fmt_perc_value_col() is an internal utility function for create a percent value column based on a specified value column and denominator column (following the name convention of the prefix + value column).

Usage

fmt_perc_value_col(
  data,
  value_col = "value",
  denominator_prefix = "denominator_",
  denominator_data = NULL,
  ...,
  perc_prefix = "perc_",
  digits = 2
)

Arguments

data

Data to use in creating the percent value column.

denominator_data

Optional data frame with prepared denominator data.

...

Additional parameters passed to dplyr::left_join.