Get IPUMS Extract histories

Get IPUMS Extract histories

Usage

get_nhgis_extract_history(...)

get_usa_extract_history(...)

get_cps_extract_history(...)

get_ipumsi_extract_history(...)

Arguments

...

Arguments passed on to ipumsr::get_extract_history

how_many

The number of extract requests for which to retrieve information. Defaults to the 10 most recent extracts.

delay

Number of seconds to delay between successive API requests, if multiple requests are needed to retrieve all records.

A delay is highly unlikely to be necessary and is intended only as a fallback in the event that you cannot retrieve your extract history without exceeding the API rate limit.

api_key

API key associated with your user account. Defaults to the value of the IPUMS_API_KEY environment variable. See set_ipums_api_key().