List NHGIS metadata

list_nhgis_metadata() uses ipumsr::get_metadata_nhgis to return and (optionally) cache a data frame with metadata for NHGIS tables.

Usage

list_nhgis_metadata(
  ...,
  cache = FALSE,
  cache_file = "nhgis_time_series_tables.rds",
  refresh = FALSE,
  integration = NULL
)

Arguments

...

Arguments passed on to ipumsr::get_metadata_nhgis

dataset

Name of an individual dataset for which to retrieve metadata.

data_table

Name of an individual data table for which to retrieve metadata. If provided, an associated dataset must also be specified.

time_series_table

Name of an individual time series table for which to retrieve metadata.

delay

Number of seconds to delay between successive API requests, if multiple requests are needed to retrieve all records.

A delay is highly unlikely to be necessary and is intended only as a fallback in the event that you cannot retrieve all metadata records without exceeding the API rate limit.

Only used if type is provided.

api_key

API key associated with your user account. Defaults to the value of the IPUMS_API_KEY environment variable. See set_ipums_api_key().

cache

If TRUE, cache metadata to file.

cache_file

Default file name to use for cached metadata.

refresh

If TRUE, force a refresh of the cached metadata file.

integration

Optionally filter to "Standardized to 2010" or "Nominal" geographic integration. Values of "2010" or "standardized" are treated as "Standardized to 2010".