Download IPUMS extract using ipumsr::wait_for_extract and ipumsr::download_extract

download_ipumsr_extract() is a wrapper for ipumsr::wait_for_extract() and ipumsr::download_extract() to wait until an extract is ready for download before attempting to download it.

Usage

download_ipumsr_extract(
  extract = NULL,
  download_dir = getwd(),
  overwrite = FALSE,
  progress = TRUE,
  ...,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

Arguments

extract

One of:

An ipums_extract object

The data collection and extract number formatted as a string of the form "collection:number" or as a vector of the form c("collection", number)

An extract number to be associated with your default IPUMS collection. See set_ipums_default_collection()

For a list of codes used to refer to each collection, see ipums_data_collections().

download_dir

Path to the directory where the files should be written. Defaults to current working directory.

overwrite

If TRUE, overwrite any conflicting files that already exist in download_dir. Defaults to FALSE.

progress

If TRUE, output progress bar showing the status of the download request. Defaults to TRUE.

...

Arguments passed on to ipumsr::wait_for_extract

initial_delay_seconds

Seconds to wait before first status check. The wait time will automatically increase by 10 seconds between each successive check.

max_delay_seconds

Maximum interval to wait between status checks. When the wait interval reaches this value, checks will continue to occur at max_delay_seconds intervals until the extract is complete or timeout_seconds is reached. Defaults to 300 seconds (5 minutes).

timeout_seconds

Maximum total number of seconds to continue waiting for the extract before throwing an error. Defaults to 10,800 seconds (3 hours).

verbose

If TRUE, print status updates to the R console at the beginning of each wait interval and upon extract completion. Defaults to TRUE.

api_key

API key associated with your user account. Defaults to the value of the IPUMS_API_KEY environment variable. See set_ipums_api_key().